requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
requests
django-debug-toolbar
pytz
google.auth.transport.requests